+++ pypeline/tools/phylo_pipeline/config.py
@@ -95,6 +95,8 @@ def _run_config_parser(argv):
     pypeline.logger.add_optiongroup(parser, default = PerHostValue("warning"))
 
     group  = optparse.OptionGroup(parser, "Scheduling")
+    group.add_option("--samtools-max-threads",  default = PerHostValue(1), type = int,
+                     help = "Maximum number of threads to use when genotyping or building pileups [%default]")
     group.add_option("--examl-max-threads",  default = PerHostValue(1), type = int,
                      help = "Maximum number of threads to use for each instance of ExaML [%default]")
     group.add_option("--max-threads",        default = per_host_cfg.max_threads, type = int,